feat: removed routes from app
