+++ playing-with-numbers/App.js
@@ -1,26 +1,12 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Login from './src/Pages/login';
-import Game from './src/Pages/game';
+import Routes from './src/routes';
 
-const Stack = createNativeStackNavigator();
 
 export default function App() {
   return (
     <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen
-          name="Login"
-          component={Login}
-          // options={{ title: 'Welcome' }}
-        />
-        <Stack.Screen
-          name="Game"
-          component={Game}
-          // options={{ title: 'Welcome' }}
-        />
-      </Stack.Navigator>
+      <Routes />
     </NavigationContainer>
   );
 }
