+++ playing-with-numbers/src/routes/index.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import Login from '../Screens/login';
+import Game from '../Screens/game';
+
+const Stack = createNativeStackNavigator();
+
+export default function Routes() {
+  return (
+    <Stack.Navigator>
+      <Stack.Screen
+        name="Login"
+        component={Login}
+      />
+      <Stack.Screen
+        name="Game"
+        component={Game}
+      />
+  </Stack.Navigator>
+  )
+}
